src/dynamic_analysis.py
# dynamic_analysis.py
import frida

def run_dynamic_analysis(package_name):
    # Attach to a running process on a device/emulator
    session = frida.get_usb_device().attach(package_name)
    script = session.create_script("""
    // Insert Frida script to monitor function calls, APIs, etc.
    """)
    script.on('message', lambda message, data: print(message))
    script.load()
    print(f"Running dynamic analysis on {package_name}")

